requirements.txt
google-generativeai
google-api-python-client
pandas
python-dotenv
tqdm
markdown
requests
yt-dlp>=2026.03.13
streamlit
tabulate
flask
gunicorn
google-cloud-run
